backend/receiveTicket/requirements.txt
functions-framework==3.*
flask
google-cloud-firestore
google-cloud-pubsub